Moved the files into folders for better orientation.
