+++ Labyrinth.Console/Controllers/IFlowController.cs
@@ -1,7 +1,12 @@
-﻿namespace Labyrinth.Console.Controllers
+﻿using Labyrinth.Console.Obstacles;
+
+namespace Labyrinth.Console.Controllers
 {
     public interface IFlowController
     {
         void RenderObstacle(Obstacle obstacle);
+
+        public void QuickObstacleGeneration(Coordinates currentCoordinates, ObstacleEdges currentEdges);
     }
+
 }
