+++ Labyrinth.Console/PlayerMovements/CalculateNewCoordinates.cs
@@ -1,11 +1,11 @@
-﻿using LabyrinthConsole;
+﻿using Labyrinth.Console;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Labyrinth.Console
+namespace Labyrinth.Console.PlayerMovements
 {
     public static class CalculateCoords
     {
